+++ app/backend/events/event_warnings.py
@@ -6,6 +6,7 @@ from app.objects.event_warnings import ListOfEventWarnings
 from app.objects.events import Event
 
 
+
 def get_list_of_event_warnings(object_store: ObjectStore, event: Event)-> ListOfEventWarnings:
     return object_store.get(
         object_definition_for_list_of_event_warnings,
@@ -39,11 +40,29 @@ def mark_event_warning_with_id_as_ignore(object_store: ObjectStore, event: Event
     update_list_of_event_warnings(list_of_event_warnings=list_of_warnings, event=event, object_store=object_store)
 
 
+def mark_event_warning_with_id_as_unignore(object_store: ObjectStore, event: Event, warning_id:str):
+    list_of_warnings = get_list_of_event_warnings(object_store=object_store, event=event)
+    list_of_warnings.mark_event_warning_with_id_as_unignored(warning_id)
+    update_list_of_event_warnings(list_of_event_warnings=list_of_warnings, event=event, object_store=object_store)
+
 def get_list_of_warnings_at_event_for_categories_sorted_by_category_and_priority(object_store: ObjectStore, event: Event, list_of_categories: List[str]) -> ListOfEventWarnings:
     list_of_warnings = get_list_of_event_warnings(object_store=object_store, event=event)
     return list_of_warnings.get_list_of_warnings_at_event_for_categories_sorted_by_priority_and_category(list_of_categories=list_of_categories)
 
 
-def get_list_of_active_warning_ids_at_event(object_store: ObjectStore, event: Event) -> list[str]:
+def get_list_of_all_warning_ids_at_event(object_store: ObjectStore, event: Event) -> list[str]:
     list_of_warnings = get_list_of_event_warnings(object_store=object_store, event=event)
-    return list_of_warnings.active_only().list_of_ids
+    return list_of_warnings.list_of_ids
+
+
+
+def mark_all_active_event_warnings_with_priority_and_category_as_ignored(object_store: ObjectStore, event: Event, category: str, priority: str):
+    list_of_warnings = get_list_of_event_warnings(object_store=object_store, event=event)
+    list_of_warnings.mark_all_active_event_warnings_with_priority_and_category_as_ignored(category=category, priority=priority)
+    update_list_of_event_warnings(list_of_event_warnings=list_of_warnings, event=event, object_store=object_store)
+
+
+def mark_all_ignored_event_warnings_with_priority_and_category_as_unignored(object_store: ObjectStore, event: Event,  category: str, priority: str):
+    list_of_warnings = get_list_of_event_warnings(object_store=object_store, event=event)
+    list_of_warnings.mark_all_ignored_event_warnings_with_priority_and_category_as_unignored(category=category, priority=priority)
+    update_list_of_event_warnings(list_of_event_warnings=list_of_warnings, event=event, object_store=object_store)