+++ app/frontend/events/volunteer_rota/ENTRY1_display_main_rota_page.py
@@ -37,7 +37,8 @@ from app.frontend.events.volunteer_rota.rota_state import (
     clear_all_filters,
 )
 from app.frontend.shared.buttons import is_button_sort_order
-from app.frontend.shared.warnings_table import save_warnings_button, save_warnings_from_table
+from app.frontend.shared.warnings_table import save_warnings_button, save_warnings_from_table, \
+    is_save_warnings_button_pressed
 from app.objects.abstract_objects.abstract_form import (
     Form,
     NewForm,
@@ -212,7 +213,7 @@ def post_form_view_for_volunteer_rota_if_data_changed(
     elif save_group_notes_button.pressed(last_button_pressed):
         save_group_notes_from_form(interface)
 
-    elif save_warnings_button.pressed(last_button_pressed):
+    elif is_save_warnings_button_pressed(last_button_pressed):
         save_warnings_from_table(interface)
 
     else: